requirements.txt
absl-py==0.1.13
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.6.2
astroid==1.6.3
astropy==3.0.2
attrs==17.4.0
Automat==0.7.0
Babel==2.5.3
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bcolz==1.2.0
beautifulsoup4==4.6.0
better-exceptions==0.1.8
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.15
boto==2.48.0
Bottleneck==1.2.1
cached-property==1.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.5.11
conda-build==3.9.2
conda-verify==2.0.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3.1
cvxopt==1.1.9
cycler==0.10.0
cypari==2.3.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.2
datashape==0.5.4
decorator==4.3.0
dill==0.2.5
distributed==1.21.6
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==0.12.2
Flask-Cors==3.0.3
funcat==0.3.2
future==0.16.0
FXrays==1.3.3
gast==0.2.0
gevent==1.2.2
glob2==0.6
greenlet==0.4.13
grpcio==1.10.1
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.11.3
hyperlink==18.0.0
idna==2.6
imageio==2.3.0
imagesize==1.0.0
incremental==17.5.0
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jsonpickle==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.0
jupyterlab-launcher==0.10.5
Keras==2.1.5
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
line-profiler==2.1.2
llvmlite==0.22.0
locket==0.2.0
Logbook==1.0.0
lru-dict==1.1.6
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.11
mistune==0.8.3
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.4.7
multipledispatch==0.5.0
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.6.0
numba==0.37.0
numexpr==2.6.2
numpy==1.14.2
numpydoc==0.8.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
pandas==0.19.2
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
plink==2.2
pluggy==0.6.0
ply==3.11
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==5.4.5
Pweave==0.30.2
py==1.5.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4
PyMemoize==1.0.2
pymongo==3.7.1
PyMySQL==0.9.2
pyodbc==4.0.23
pyOpenSSL==17.5.0
pyparsing==2.2.0
pypng==0.0.18
pyreadline==2.1
PySocks==1.6.8
pytest==3.5.0
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.2
pytz==2018.4
PyWavelets==0.5.2
pywin32==223
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.5.0
redis==2.10.6
requests==2.18.4
rope==0.10.7
rqalpha-mod-ricequant-data==2.0.6.15.dev1
rqdatac==0.4.1
rqportal-internal==1.0.16a0
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8.1
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
snappy==2.6
snappy-manifolds==1.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
spherogram==1.8
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.3.1
spyder-kernels==0.2.6
spyder-reports==0.1.1
SQLAlchemy==1.2.11
statsmodels==0.8.0
sympy==1.1.1
TA-Lib==0.4.17
tables==3.4.3
tabulate==0.8.2
tblib==1.3.2
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.0.2
traitlets==4.3.2
tushare==1.2.15
Twisted==18.7.0
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.7
xlwt==1.3.0
zict==0.1.3
zope.interface==4.5.0
